files renaming compilation fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@5739 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Tlen/src/tlen_thread.cpp
@@ -21,11 +21,11 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 */
 
-#include "jabber.h"
+#include "tlen.h"
 
 #include "commons.h"
-#include "jabber_list.h"
-#include "jabber_iq.h"
+#include "tlen_list.h"
+#include "tlen_iq.h"
 #include "resource.h"
 #include "tlen_p2p_old.h"
 #include "tlen_file.h"
@@ -624,7 +624,7 @@ static void JabberProcessMessage(XmlNode *node, ThreadData *info)
 				if ((bodyNode=JabberXmlGetChild(node, "body")) != NULL) {
 					if (bodyNode->text != NULL) {
 						if ((subjectNode=JabberXmlGetChild(node, "subject")) != NULL && subjectNode->text != NULL && subjectNode->text[0] != '\0') {
-							int size = strlen(subjectNode->text)+strlen(bodyNode->text)+5;
+							size_t size = strlen(subjectNode->text)+strlen(bodyNode->text)+5;
 							p = (char *)mir_alloc(size);
 							mir_snprintf(p, size, "%s\r\n%s", subjectNode->text, bodyNode->text);
 							localMessage = JabberTextDecode(p);
